+++ src/05-Params.ps1
@@ -19,6 +19,9 @@ param(
 
     [int]$EventLogDays = 0,
 
+    [switch]$SkipUpdateSearch,
+    [int]$UpdateSearchTimeoutSec = 180,
+
     [string]$EmailTo,
     [string]$EmailFrom = "systemaudit@$($env:COMPUTERNAME).local",
     [string]$SmtpServer = ''